Fullfør Infopanel-migrering
