+++ src/pages/Sykmeldingside/sykmeldinger/AvbruttSykmelding/AvbruttSykmelding.tsx
@@ -1,43 +1,46 @@
 import React from 'react';
-import { Sykmelding } from '../../../types/sykmeldingTypes';
-import Sidetopp from '../components/sidetopp/Sidetopp';
-import SykmeldingPerioder from '../components/Infopanel/panelelementer/periode/SykmeldingPerioder';
-import DiagnoseSeksjon from '../components/Infopanel/panelelementer/diagnose/DiagnoseSeksjon';
-import FraverSeksjon from '../components/Infopanel/panelelementer/FraverSeksjon';
-import SvangerskapSeksjon from '../components/Infopanel/panelelementer/SvangerskapSeksjon';
-import SkadeSeksjon from '../components/Infopanel/panelelementer/SkadeSeksjon';
-import PrognoseSeksjon from '../components/Infopanel/panelelementer/PrognoseSeksjon';
-import ArbeidsuforSeksjon from '../components/Infopanel/panelelementer/ArbeidsuforSeksjon';
-import ArbeidsgiverSeksjon from '../components/Infopanel/panelelementer/ArbeidsgiverSeksjon';
-import LegeSeksjon from '../components/Infopanel/panelelementer/LegeSeksjon';
-import Utvidbar from '../../../components/utvidbar/Utvidbar';
-import BehandlingsDatoer from '../components/Infopanel/utdypendeelementer/BehandlingsDatoer';
-import MulighetForArbeid from '../components/Infopanel/utdypendeelementer/MulighetForArbeid';
-import Friskmelding from '../components/Infopanel/utdypendeelementer/Friskmelding';
-import UtdypendeOpplysninger from '../components/Infopanel/utdypendeelementer/UtdypendeOpplysninger';
-import Arbeidsevne from '../components/Infopanel/utdypendeelementer/Arbeidsevne';
-import SeksjonMedTittel from '../components/Infopanel/layout/SeksjonMedTittel';
-import ElementMedTekst from '../components/Infopanel/layout/ElementMedTekst';
+
+import { Sykmelding } from '../../../../types/sykmeldingTypes';
+
+import Sidetopp from '../../components/Sidetopp/Sidetopp';
+import Utvidbar from '../../../../components/utvidbar/Utvidbar';
+import AvbruttStatuspanel from '../../components/Statuspanel/AvbruttStatuspanel';
+
+import SykmeldingPerioder from '../../components/Infopanel/panelelementer/periode/SykmeldingPerioder';
+import DiagnoseSeksjon from '../../components/Infopanel/panelelementer/diagnose/DiagnoseSeksjon';
+import FraverSeksjon from '../../components/Infopanel/panelelementer/FraverSeksjon';
+import SvangerskapSeksjon from '../../components/Infopanel/panelelementer/SvangerskapSeksjon';
+import SkadeSeksjon from '../../components/Infopanel/panelelementer/SkadeSeksjon';
+import PrognoseSeksjon from '../../components/Infopanel/panelelementer/PrognoseSeksjon';
+import ArbeidsuforSeksjon from '../../components/Infopanel/panelelementer/ArbeidsuforSeksjon';
+import ArbeidsgiverSeksjon from '../../components/Infopanel/panelelementer/ArbeidsgiverSeksjon';
+import LegeSeksjon from '../../components/Infopanel/panelelementer/LegeSeksjon';
+import BehandlingsDatoer from '../../components/Infopanel/utdypendeelementer/BehandlingsDatoer';
+import MulighetForArbeid from '../../components/Infopanel/utdypendeelementer/MulighetForArbeid';
+import Friskmelding from '../../components/Infopanel/utdypendeelementer/Friskmelding';
+import UtdypendeOpplysninger from '../../components/Infopanel/utdypendeelementer/UtdypendeOpplysninger';
+import Arbeidsevne from '../../components/Infopanel/utdypendeelementer/Arbeidsevne';
+import SeksjonMedTittel from '../../components/Infopanel/layout/SeksjonMedTittel';
+import ElementMedTekst from '../../components/Infopanel/layout/ElementMedTekst';
 
 import doktor from '../../../svg/doktor.svg';
 import doktorHover from '../../../svg/doktorHover.svg';
 import person from '../../../svg/person.svg';
 import personHover from '../../../svg/personHover.svg';
 
-import tekster from './BekreftetSykmelding-tekster';
-import BekreftetStatuspanel from '../components/Statuspanel/BekreftetStatuspanel';
-import Tittel from '../components/Infopanel/layout/Tittel';
+import tekster from './AvbruttSykmelding-tekster';
+import Tittel from '../../components/Infopanel/layout/Tittel';
 
-interface BekreftetSykmeldingProps {
+interface AvbruttSykmeldingProps {
     sykmelding: Sykmelding;
 }
 
-const BekreftetSykmelding = ({ sykmelding }: BekreftetSykmeldingProps) => {
+const AvbruttSykmelding = ({ sykmelding }: AvbruttSykmeldingProps) => {
     return (
         <div className="sykmelding-container">
             <Sidetopp tekst="Sykmelding" />
 
-            <BekreftetStatuspanel sykmelding={sykmelding} />
+            <AvbruttStatuspanel sykmelding={sykmelding} />
 
             <Utvidbar apen tittel="Dine opplysninger" fargetema="info" ikon={person} ikonHover={personHover}>
                 <Tittel tekst="Sykmelding" />
@@ -57,7 +60,7 @@ const BekreftetSykmelding = ({ sykmelding }: BekreftetSykmeldingProps) => {
                 <Utvidbar
                     ikon={doktor}
                     ikonHover={doktorHover}
-                    tittel={tekster['bekreftet-sykmelding.flere-opplysninger.tittel']}
+                    tittel={tekster['avbrutt-sykmelding.flere-opplysninger.tittel']}
                 >
                     <BehandlingsDatoer
                         behandletTidspunkt={sykmelding.behandletTidspunkt}
@@ -79,4 +82,4 @@ const BekreftetSykmelding = ({ sykmelding }: BekreftetSykmeldingProps) => {
     );
 };
 
-export default BekreftetSykmelding;
+export default AvbruttSykmelding;
