+++ src/pages/Sykmeldingside/sykmeldinger/AvvistSykmelding/components/Begrunnelser.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Element } from 'nav-frontend-typografi';
 
 import tekster from '../AvvistSykmelding-tekster';
-import { Sykmelding } from '../../../../types/sykmeldingTypes';
+import { Sykmelding } from '../../../../../types/sykmeldingTypes';
 
 export enum Avvisningsregler {
     PASIENT_ELDRE_ENN_70 = 'PASIENT_ELDRE_ENN_70',
