+++ src/pages/Sykmeldingside/sykmeldinger/AvvistSykmelding/components/veilederinnhold.less
