+++ src/pages/Sykmeldingside/sykmeldinger/BekreftetSykmelding/BekreftetSykmelding.tsx
@@ -1,46 +1,43 @@
 import React from 'react';
-
-import { Sykmelding } from '../../../types/sykmeldingTypes';
-
-import Sidetopp from '../components/Sidetopp/Sidetopp';
-import Utvidbar from '../../../components/utvidbar/Utvidbar';
-import AvbruttStatuspanel from '../components/Statuspanel/AvbruttStatuspanel';
-
-import SykmeldingPerioder from '../components/Infopanel/panelelementer/periode/SykmeldingPerioder';
-import DiagnoseSeksjon from '../components/Infopanel/panelelementer/diagnose/DiagnoseSeksjon';
-import FraverSeksjon from '../components/Infopanel/panelelementer/FraverSeksjon';
-import SvangerskapSeksjon from '../components/Infopanel/panelelementer/SvangerskapSeksjon';
-import SkadeSeksjon from '../components/Infopanel/panelelementer/SkadeSeksjon';
-import PrognoseSeksjon from '../components/Infopanel/panelelementer/PrognoseSeksjon';
-import ArbeidsuforSeksjon from '../components/Infopanel/panelelementer/ArbeidsuforSeksjon';
-import ArbeidsgiverSeksjon from '../components/Infopanel/panelelementer/ArbeidsgiverSeksjon';
-import LegeSeksjon from '../components/Infopanel/panelelementer/LegeSeksjon';
-import BehandlingsDatoer from '../components/Infopanel/utdypendeelementer/BehandlingsDatoer';
-import MulighetForArbeid from '../components/Infopanel/utdypendeelementer/MulighetForArbeid';
-import Friskmelding from '../components/Infopanel/utdypendeelementer/Friskmelding';
-import UtdypendeOpplysninger from '../components/Infopanel/utdypendeelementer/UtdypendeOpplysninger';
-import Arbeidsevne from '../components/Infopanel/utdypendeelementer/Arbeidsevne';
-import SeksjonMedTittel from '../components/Infopanel/layout/SeksjonMedTittel';
-import ElementMedTekst from '../components/Infopanel/layout/ElementMedTekst';
+import { Sykmelding } from '../../../../types/sykmeldingTypes';
+import Sidetopp from '../../components/sidetopp/Sidetopp';
+import SykmeldingPerioder from '../../components/Infopanel/panelelementer/periode/SykmeldingPerioder';
+import DiagnoseSeksjon from '../../components/Infopanel/panelelementer/diagnose/DiagnoseSeksjon';
+import FraverSeksjon from '../../components/Infopanel/panelelementer/FraverSeksjon';
+import SvangerskapSeksjon from '../../components/Infopanel/panelelementer/SvangerskapSeksjon';
+import SkadeSeksjon from '../../components/Infopanel/panelelementer/SkadeSeksjon';
+import PrognoseSeksjon from '../../components/Infopanel/panelelementer/PrognoseSeksjon';
+import ArbeidsuforSeksjon from '../../components/Infopanel/panelelementer/ArbeidsuforSeksjon';
+import ArbeidsgiverSeksjon from '../../components/Infopanel/panelelementer/ArbeidsgiverSeksjon';
+import LegeSeksjon from '../../components/Infopanel/panelelementer/LegeSeksjon';
+import Utvidbar from '../../../../components/utvidbar/Utvidbar';
+import BehandlingsDatoer from '../../components/Infopanel/utdypendeelementer/BehandlingsDatoer';
+import MulighetForArbeid from '../../components/Infopanel/utdypendeelementer/MulighetForArbeid';
+import Friskmelding from '../../components/Infopanel/utdypendeelementer/Friskmelding';
+import UtdypendeOpplysninger from '../../components/Infopanel/utdypendeelementer/UtdypendeOpplysninger';
+import Arbeidsevne from '../../components/Infopanel/utdypendeelementer/Arbeidsevne';
+import SeksjonMedTittel from '../../components/Infopanel/layout/SeksjonMedTittel';
+import ElementMedTekst from '../../components/Infopanel/layout/ElementMedTekst';
 
 import doktor from '../../../svg/doktor.svg';
 import doktorHover from '../../../svg/doktorHover.svg';
 import person from '../../../svg/person.svg';
 import personHover from '../../../svg/personHover.svg';
 
-import tekster from './AvbruttSykmelding-tekster';
-import Tittel from '../components/Infopanel/layout/Tittel';
+import tekster from './BekreftetSykmelding-tekster';
+import BekreftetStatuspanel from '../../components/Statuspanel/BekreftetStatuspanel';
+import Tittel from '../../components/Infopanel/layout/Tittel';
 
-interface AvbruttSykmeldingProps {
+interface BekreftetSykmeldingProps {
     sykmelding: Sykmelding;
 }
 
-const AvbruttSykmelding = ({ sykmelding }: AvbruttSykmeldingProps) => {
+const BekreftetSykmelding = ({ sykmelding }: BekreftetSykmeldingProps) => {
     return (
         <div className="sykmelding-container">
             <Sidetopp tekst="Sykmelding" />
 
-            <AvbruttStatuspanel sykmelding={sykmelding} />
+            <BekreftetStatuspanel sykmelding={sykmelding} />
 
             <Utvidbar apen tittel="Dine opplysninger" fargetema="info" ikon={person} ikonHover={personHover}>
                 <Tittel tekst="Sykmelding" />
@@ -60,7 +57,7 @@ const AvbruttSykmelding = ({ sykmelding }: AvbruttSykmeldingProps) => {
                 <Utvidbar
                     ikon={doktor}
                     ikonHover={doktorHover}
-                    tittel={tekster['avbrutt-sykmelding.flere-opplysninger.tittel']}
+                    tittel={tekster['bekreftet-sykmelding.flere-opplysninger.tittel']}
                 >
                     <BehandlingsDatoer
                         behandletTidspunkt={sykmelding.behandletTidspunkt}
@@ -82,4 +79,4 @@ const AvbruttSykmelding = ({ sykmelding }: AvbruttSykmeldingProps) => {
     );
 };
 
-export default AvbruttSykmelding;
+export default BekreftetSykmelding;
