+++ src/pages/Sykmeldingside/sykmeldinger/NySykmelding/NySykmelding.tsx
@@ -1,32 +1,32 @@
 import React, { useRef, useEffect } from 'react';
 import { Normaltekst } from 'nav-frontend-typografi';
-import Tittel from '../components/Infopanel/layout/Tittel';
-import { Sykmelding } from '../../../types/sykmeldingTypes';
+import Tittel from '../../components/Infopanel/layout/Tittel';
+import { Sykmelding } from '../../../../types/sykmeldingTypes';
 import tekster from './NySykmelding-tekster';
-import Sidetopp from '../components/sidetopp/Sidetopp';
-import Infopanel from '../components/Infopanel/Infopanel';
-import Veileder from '../../../components/veileder/Veileder';
+import Sidetopp from '../../components/sidetopp/Sidetopp';
+import Infopanel from '../../components/Infopanel/Infopanel';
+import Veileder from '../../../../components/veileder/Veileder';
 import EldreSykmeldingVarsel from './components/EldreSykmeldingVarsel';
-import SykmeldingPerioder from '../components/Infopanel/panelelementer/periode/SykmeldingPerioder';
-import DiagnoseSeksjon from '../components/Infopanel/panelelementer/diagnose/DiagnoseSeksjon';
-import FraverSeksjon from '../components/Infopanel/panelelementer/FraverSeksjon';
-import SvangerskapSeksjon from '../components/Infopanel/panelelementer/SvangerskapSeksjon';
-import SkadeSeksjon from '../components/Infopanel/panelelementer/SkadeSeksjon';
-import ArbeidsuforSeksjon from '../components/Infopanel/panelelementer/ArbeidsuforSeksjon';
-import PrognoseSeksjon from '../components/Infopanel/panelelementer/PrognoseSeksjon';
-import ArbeidsgiverSeksjon from '../components/Infopanel/panelelementer/ArbeidsgiverSeksjon';
-import LegeSeksjon from '../components/Infopanel/panelelementer/LegeSeksjon';
-import BehandlingsDatoer from '../components/Infopanel/utdypendeelementer/BehandlingsDatoer';
-import MulighetForArbeid from '../components/Infopanel/utdypendeelementer/MulighetForArbeid';
-import Friskmelding from '../components/Infopanel/utdypendeelementer/Friskmelding';
-import UtdypendeOpplysninger from '../components/Infopanel/utdypendeelementer/UtdypendeOpplysninger';
-import Arbeidsevne from '../components/Infopanel/utdypendeelementer/Arbeidsevne';
-import SeksjonMedTittel from '../components/Infopanel/layout/SeksjonMedTittel';
-import ElementMedTekst from '../components/Infopanel/layout/ElementMedTekst';
-import Sporsmal from '../../../components/sporsmal/Sporsmal';
-import SporsmalInfoheader from '../../../components/sporsmal/SporsmalInfoheader';
-import Arbeidsgiver from '../../../types/arbeidsgiverTypes';
-import Utvidbar from '../../../components/utvidbar/Utvidbar';
+import SykmeldingPerioder from '../../components/Infopanel/panelelementer/periode/SykmeldingPerioder';
+import DiagnoseSeksjon from '../../components/Infopanel/panelelementer/diagnose/DiagnoseSeksjon';
+import FraverSeksjon from '../../components/Infopanel/panelelementer/FraverSeksjon';
+import SvangerskapSeksjon from '../../components/Infopanel/panelelementer/SvangerskapSeksjon';
+import SkadeSeksjon from '../../components/Infopanel/panelelementer/SkadeSeksjon';
+import ArbeidsuforSeksjon from '../../components/Infopanel/panelelementer/ArbeidsuforSeksjon';
+import PrognoseSeksjon from '../../components/Infopanel/panelelementer/PrognoseSeksjon';
+import ArbeidsgiverSeksjon from '../../components/Infopanel/panelelementer/ArbeidsgiverSeksjon';
+import LegeSeksjon from '../../components/Infopanel/panelelementer/LegeSeksjon';
+import BehandlingsDatoer from '../../components/Infopanel/utdypendeelementer/BehandlingsDatoer';
+import MulighetForArbeid from '../../components/Infopanel/utdypendeelementer/MulighetForArbeid';
+import Friskmelding from '../../components/Infopanel/utdypendeelementer/Friskmelding';
+import UtdypendeOpplysninger from '../../components/Infopanel/utdypendeelementer/UtdypendeOpplysninger';
+import Arbeidsevne from '../../components/Infopanel/utdypendeelementer/Arbeidsevne';
+import SeksjonMedTittel from '../../components/Infopanel/layout/SeksjonMedTittel';
+import ElementMedTekst from '../../components/Infopanel/layout/ElementMedTekst';
+import Sporsmal from '../../../../components/sporsmal/Sporsmal';
+import SporsmalInfoheader from '../../../../components/sporsmal/SporsmalInfoheader';
+import Arbeidsgiver from '../../../../types/arbeidsgiverTypes';
+import Utvidbar from '../../../../components/utvidbar/Utvidbar';
 
 import doktor from '../../../svg/doktor.svg';
 import doktorHover from '../../../svg/doktorHover.svg';
