+++ src/pages/Sykmeldingside/sykmeldinger/NySykmelding/components/EldreSykmeldingVarsel.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import tekster from '../NySykmelding-tekster';
 import Lenke from 'nav-frontend-lenker';
 import AlertStripe from 'nav-frontend-alertstriper';
-import Margin from '../../components/Infopanel/layout/Margin';
+import Margin from '../../../components/Infopanel/layout/Margin';
 
 const EldreSykmeldingVarsel = () => {
     // TODO: hvis brukeren har en eldre sykmelding som må behandles først
